Factor out match history controller code in a shared singleton for access from the diffable data source
